backend/requirements.txt
Django==4.2.4
djangorestframework==3.14.0
drf-extra-fields==3.7.0
pillow==10.0.0
PyJWT==2.8.0
psycopg2-binary==2.9.9
pytest==6.2.5
pytest-django==4.5.2
pytest-pythonpath==0.7.4
djoser==2.2.0
django-cors-headers==4.2.0
django-filter==23.3